Review fixes: DRY deduplication, dead code removal, integration
- Extract shared fetchAllClaims() into claims-data.ts (was duplicated in 2 files)
- Use ClaimStatsResult/GetClaimsResult from api-types instead of local copies
- Remove unused useClaimFilters hook from claims-filters.tsx
- Fix React Fragment key bug in claims-table.tsx (key on Fragment, not child)
- Remove dead meta.hideBelow that was never read
- Delete dead claims-dashboard.tsx (702 lines, replaced by /claims section)
- Add Claims link to wiki page metadata bar (/wiki/[id])
